refactor: move HandleRevoke to pkg/revoke, ValidateBearer uses token.TokenByAccessToken
- Extracts HandleRevoke from pkg/token into a dedicated pkg/revoke package,
  matching the existing one-endpoint-per-package convention (pkg/introspect,
  pkg/userinfo, pkg/authorize, pkg/login, pkg/signup, ...). pkg/token becomes
  a pure token-domain package again.

- Adds token.RevokeByTokenValue for the UPDATE side, keeping tokens-table
  writes in pkg/token.

- With pkg/token no longer importing pkg/bearer, the bearer → token edge is
  unblocked. ValidateBearer now uses token.TokenByAccessToken instead of
  inline SELECT — consistent with admin_auth and the rest of the tokens-table
  access pattern.

- Moves GetUserFromRequest from pkg/user to pkg/bearer (renamed
  UserFromRequest). The old location forced user → bearer, which cycled
  once bearer → token was added. UserFromRequest now lives next to its
  counterpart ValidateBearer. Audit-log callers in pkg/user/handler.go
  switched to audit.ActorFromRequest which already handled that case.

- Moves the corresponding tests (revoke handler + UserFromRequest) to the
  packages that own those functions. pkg/token/handler_test.go loses four
  tests whose coverage duplicates what's now in pkg/revoke/handler_test.go.
