+++ pkg/account/trusted_devices.go
@@ -3,8 +3,8 @@ package account
 import (
 	"net/http"
 
+	"github.com/eugenioenko/autentico/pkg/bearer"
 	"github.com/eugenioenko/autentico/pkg/trusteddevice"
-	"github.com/eugenioenko/autentico/pkg/user"
 	"github.com/eugenioenko/autentico/pkg/utils"
 )
 
@@ -18,7 +18,7 @@ import (
 // @Failure 401 {object} model.ApiError
 // @Router /account/api/trusted-devices [get]
 func HandleListTrustedDevices(w http.ResponseWriter, r *http.Request) {
-	usr, err := user.GetUserFromRequest(r)
+	usr, err := bearer.UserFromRequest(r)
 	if err != nil {
 		utils.WriteErrorResponse(w, http.StatusUnauthorized, "unauthorized", err.Error())
 		return
@@ -57,7 +57,7 @@ func HandleListTrustedDevices(w http.ResponseWriter, r *http.Request) {
 // @Failure 403 {object} model.ApiError
 // @Router /account/api/trusted-devices/{id} [delete]
 func HandleRevokeTrustedDevice(w http.ResponseWriter, r *http.Request) {
-	usr, err := user.GetUserFromRequest(r)
+	usr, err := bearer.UserFromRequest(r)
 	if err != nil {
 		utils.WriteErrorResponse(w, http.StatusUnauthorized, "unauthorized", err.Error())
 		return
