+++ pkg/cli/start.go
@@ -40,6 +40,7 @@ import (
 	"github.com/eugenioenko/autentico/pkg/passwordreset"
 	"github.com/eugenioenko/autentico/pkg/ratelimit"
 	"github.com/eugenioenko/autentico/pkg/reqid"
+	"github.com/eugenioenko/autentico/pkg/revoke"
 	"github.com/eugenioenko/autentico/pkg/session"
 	"github.com/eugenioenko/autentico/pkg/signup"
 	"github.com/eugenioenko/autentico/pkg/token"
@@ -148,7 +149,7 @@ func RunStart(c *cli.Context) error {
 	mux.Handle(oauth+"/signup/", csrfProtected(signup.HandleSignup))
 	mux.Handle("POST "+oauth+"/token", rateLimitedFunc(token.HandleToken))
 	mux.Handle("POST "+oauth+"/protocol/openid-connect/token", rateLimitedFunc(token.HandleToken))
-	mux.HandleFunc("POST "+oauth+"/revoke", token.HandleRevoke)
+	mux.HandleFunc("POST "+oauth+"/revoke", revoke.HandleRevoke)
 	mux.HandleFunc("POST "+oauth+"/introspect", introspect.HandleIntrospect)
 	mux.HandleFunc(oauth+"/userinfo", userinfo.HandleUserInfo)
 	mux.HandleFunc(oauth+"/protocol/openid-connect/userinfo", userinfo.HandleUserInfo)
