+++ pkg/user/handler_test.go
@@ -60,122 +60,6 @@ func setupAuthenticatedUser(t *testing.T) (string, string) {
 	return signedToken, userID
 }
 
-// --- GetUserFromRequest tests ---
-
-func TestGetUserFromRequest_MissingAuth(t *testing.T) {
-	testutils.WithTestDB(t)
-
-	req := httptest.NewRequest(http.MethodGet, "/", nil)
-	_, err := GetUserFromRequest(req)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "missing Authorization header")
-}
-
-func TestGetUserFromRequest_InvalidAuthFormat(t *testing.T) {
-	testutils.WithTestDB(t)
-
-	req := httptest.NewRequest(http.MethodGet, "/", nil)
-	req.Header.Set("Authorization", "Basic dXNlcjpwYXNz")
-	_, err := GetUserFromRequest(req)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "invalid Authorization header")
-}
-
-func TestGetUserFromRequest_InvalidToken(t *testing.T) {
-	testutils.WithTestDB(t)
-
-	req := httptest.NewRequest(http.MethodGet, "/", nil)
-	req.Header.Set("Authorization", "Bearer invalid-token")
-	_, err := GetUserFromRequest(req)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "invalid token")
-}
-
-func TestGetUserFromRequest_NoSession(t *testing.T) {
-	testutils.WithTestDB(t)
-
-	// Generate a valid JWT but don't create a session
-	userID := xid.New().String()
-	sessionID := xid.New().String()
-	accessTokenExpiresAt := time.Now().Add(config.Get().AuthAccessTokenExpiration).UTC()
-
-	accessClaims := jwt.MapClaims{
-		"exp":   accessTokenExpiresAt.Unix(),
-		"iat":   time.Now().Unix(),
-		"iss":   config.GetBootstrap().AppAuthIssuer,
-		"aud":   config.Get().AuthAccessTokenAudience,
-		"sub":   userID,
-		"typ":   "Bearer",
-		"sid":   sessionID,
-		"scope": "openid profile email",
-	}
-	accessToken := jwt.NewWithClaims(jwt.SigningMethodRS256, accessClaims)
-	accessToken.Header["kid"] = config.GetBootstrap().AuthJwkCertKeyID
-	signedToken, err := accessToken.SignedString(key.GetPrivateKey())
-	assert.NoError(t, err)
-
-	req := httptest.NewRequest(http.MethodGet, "/", nil)
-	req.Header.Set("Authorization", "Bearer "+signedToken)
-	_, err = GetUserFromRequest(req)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "invalid session")
-}
-
-func TestGetUserFromRequest_Valid(t *testing.T) {
-	testutils.WithTestDB(t)
-
-	token, _ := setupAuthenticatedUser(t)
-	req := httptest.NewRequest(http.MethodGet, "/", nil)
-	req.Header.Set("Authorization", "Bearer "+token)
-	user, err := GetUserFromRequest(req)
-	assert.NoError(t, err)
-	assert.NotNil(t, user)
-}
-
-// Regression for https://github.com/eugenioenko/autentico/issues/225:
-// a token whose session has been deactivated (self-service logout or admin
-// revocation) must not be accepted by the account API.
-func TestGetUserFromRequest_DeactivatedSession(t *testing.T) {
-	testutils.WithTestDB(t)
-
-	token, _ := setupAuthenticatedUser(t)
-	_, err := db.GetDB().Exec(
-		`UPDATE sessions SET deactivated_at = CURRENT_TIMESTAMP WHERE access_token = ?`,
-		token,
-	)
-	assert.NoError(t, err)
-
-	req := httptest.NewRequest(http.MethodGet, "/", nil)
-	req.Header.Set("Authorization", "Bearer "+token)
-	_, err = GetUserFromRequest(req)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "deactivated")
-}
-
-// A token revoked via /oauth2/revoke (RFC 7009 — sets tokens.revoked_at)
-// must also be rejected by the account API, same class of bug as #225 but
-// on the tokens table instead of sessions.
-func TestGetUserFromRequest_RevokedToken(t *testing.T) {
-	testutils.WithTestDB(t)
-
-	token, userID := setupAuthenticatedUser(t)
-	// Persist a tokens row so /oauth2/revoke's revoked_at setter has
-	// something to mark, matching the real-world flow.
-	now := time.Now().UTC()
-	_, err := db.GetDB().Exec(`
-		INSERT INTO tokens (id, user_id, access_token, refresh_token, access_token_type,
-			refresh_token_expires_at, access_token_expires_at, issued_at, scope, grant_type, revoked_at)
-		VALUES (?, ?, ?, 'refresh', 'Bearer', ?, ?, ?, 'openid', 'password', ?)
-	`, "tok-"+userID[:6], userID, token, now.Add(time.Hour), now.Add(time.Hour), now, now)
-	assert.NoError(t, err)
-
-	req := httptest.NewRequest(http.MethodGet, "/", nil)
-	req.Header.Set("Authorization", "Bearer "+token)
-	_, err = GetUserFromRequest(req)
-	assert.Error(t, err)
-	assert.Contains(t, err.Error(), "revoked")
-}
-
 // --- HandleCreateUser tests ---
 
 func TestHandleCreateUser_InvalidBody(t *testing.T) {
